move create ingestion action and call command handler
