+++ api/src/Infrastructure/School/school.module.ts
@@ -78,13 +78,10 @@ import { GetShootingByIdQueryHandler } from 'src/Application/School/Query/Shooti
 import { GetShootingAction } from './Action/Shooting/GetShootingAction';
 import { UpdateShootingAction } from './Action/Shooting/UpdateShootingAction';
 import { UpdateShootingCommandHandler } from 'src/Application/School/Command/Shooting/UpdateShootingCommandHandler';
-import { IngestionModule } from '../Ingestion/ingestion.module';
-import { CreateFileIngestionAction } from './Action/Photo/CreateFileIngestionAction';
 
 @Module({
   imports: [
     BusModule,
-    IngestionModule,
     TypeOrmModule.forFeature([
       School,
       Photo,
@@ -122,7 +119,6 @@ import { CreateFileIngestionAction } from './Action/Photo/CreateFileIngestionAct
     CreateDiscountAction,
     CountSchoolDiscountsAction,
     RemoveDiscountAction,
-    CreateFileIngestionAction,
   ],
   providers: [
     { provide: 'ICodeGenerator', useClass: CodeGeneratorAdapter },
